# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Jungle Tetris API
  description: |
    실시간 대전 및 랭킹 시스템이 포함된 웹 기반 테트리스 게임 API

    ## 실시간 통신
    이 API는 HTTP REST API와 함께 **Socket.IO**를 사용하여 실시간 기능을 제공합니다:
    - 실시간 방 관리 (입장/퇴장 알림)
    - 게임 중 실시간 점수 동기화
    - 게임 시작/종료 이벤트 (대전 모드는 Socket.IO를 통해서만 처리)
    - 연결 상태 관리

    ### Socket.IO 엔드포인트
    - **개발 서버**: `ws://localhost:8000`
    - **프로덕션 서버**: `wss://api.jungletetris.com`

    ### 주요 Socket.IO 이벤트
    - `room:join` - 방 참가 (JWT 토큰 필요)
    - `room:leave` - 방 나가기  
    - `room:player_joined` - 플레이어 참가 알림
    - `game:start` - 게임 시작 (2명 참가 시 자동)
    - `game:score_update` - 점수 업데이트
    - `game:end` - 게임 종료 (모든 플레이어 완료 시)
    - `game:waiting` - 상대방 대기 중
    - `game:disconnect` - 연결 끊김 처리

    ## 대전 게임 로직
    **중요**: 대전 게임의 시작/종료 로직은 REST API에서 Socket.IO 이벤트로 완전히 이동되었습니다:
    - **게임 시작**: 방에 2명이 참가하면 `game:start` 이벤트가 자동으로 발생
    - **게임 종료**: 모든 플레이어가 게임을 완료하면 `game:end` 이벤트 발생
    - **실시간 점수**: `game:score_update` 이벤트를 통한 실시간 점수 동기화

    이전의 `/api/game/versus/start`와 `/api/game/versus/end` 엔드포인트는 제거되었습니다.

    **📖 자세한 Socket.IO API 사양은 `websocket-api.md` 문서를 참고하세요.**
  version: 1.3.0
  contact:
    name: Jungle Tetris Team
    email: "support@jungletetris.com"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: 개발 서버
  - url: https://api.jungletetris.com
    description: 프로덕션 서버

security:
  - BearerAuth: []

paths:
  # 방 생성 API
  /api/rooms/create:
    post:
      tags:
        - Rooms
      summary: 방 생성
      description: |
        1대1 대전을 위한 게임 룸을 생성합니다.

        **인증**: JWT 토큰에서 사용자 정보를 추출하여 방을 생성합니다.

        **📡 실시간 연동**: 방 생성 후 Socket.IO를 통해 실시간 방 관리가 가능합니다.
      responses:
        "201":
          description: 방 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  room_id:
                    type: integer
                    example: 123
                  host_id:
                    type: string
                    example: "user123"
                  status:
                    type: string
                    enum: [waiting, playing, finished]
                    example: "waiting"
                  players:
                    type: array
                    items:
                      type: string
                    example: ["홍길동"]
                  created_at:
                    type: string
                    format: date-time
                    example: "2025-07-08T16:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/rooms/join:
    post:
      tags:
        - Rooms
      summary: 방 참가
      description: |
        방 번호를 입력하여 방에 참가합니다.

        **📡 실시간 연동**: 방 참가 후 Socket.IO `room:join` 이벤트를 통해 실시간 상호작용이 시작됩니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - room_id
              properties:
                room_id:
                  type: integer
                  description: 참가할 방 번호
                  example: 123
      responses:
        "200":
          description: 방 참가 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "방에 참가했습니다"
                  room_id:
                    type: integer
                    example: 123
                  players:
                    type: array
                    items:
                      type: string
                    example: ["홍길동", "김철수"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/rooms/{id}:
    delete:
      tags:
        - Rooms
      summary: 방 삭제
      description: 방장이 방을 삭제합니다
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 방 ID
          example: 123
      responses:
        "200":
          description: 방 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "방이 삭제되었습니다"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # 게임 API
  /api/game/solo/start:
    post:
      tags:
        - Game
      summary: 솔로 게임 시작
      description: 단독 플레이 모드의 테트리스 게임을 시작합니다
      responses:
        "200":
          description: 솔로 게임 시작 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  game_time:
                    type: integer
                    description: 게임 시간 (초)
                    example: 60
                  message:
                    type: string
                    example: "솔로 게임이 시작되었습니다"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/game/solo/end:
    post:
      tags:
        - Game
      summary: 솔로 게임 종료
      description: 솔로 게임의 최종 결과를 서버에 전송합니다
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - score
              properties:
                score:
                  type: integer
                  minimum: 0
                  description: 획득 점수
                  example: 2500
      responses:
        "200":
          description: 솔로 게임 종료 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "게임 결과가 저장되었습니다"
                  score:
                    type: integer
                    example: 2500
                  is_best:
                    type: boolean
                    description: 개인 최고 기록 여부
                    example: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/game/multi/start:
    post:
      tags:
        - Game
      summary: 멀티 게임 시작
      description: 방장이 멀티플레이 게임을 시작합니다. (JWT 인증 필요)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - room_id
              properties:
                room_id:
                  type: integer
                  description: 시작할 방 번호
                  example: 123
      responses:
        "200":
          description: 멀티 게임 시작 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  room_id:
                    type: integer
                    example: 123
                  game_time:
                    type: integer
                    description: 게임 시간 (초)
                    example: 60
                  players:
                    type: array
                    description: 참가자 목록
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          example: "홍길동"
                        user_id:
                          type: string
                          example: "user123"
                        score:
                          type: integer
                          example: 0
                  message:
                    type: string
                    example: "멀티 게임이 시작되었습니다"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/game/multi/end:
    post:
      tags:
        - Game
      summary: 멀티 게임 종료
      description: 멀티플레이 게임 종료 및 결과 제출 (JWT 인증 필요)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - room_id
                - score
              properties:
                room_id:
                  type: integer
                  description: 종료할 방 번호
                  example: 123
                score:
                  type: integer
                  minimum: 0
                  description: 획득 점수
                  example: 2500
      responses:
        "200":
          description: 멀티 게임 종료 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  room_id:
                    type: integer
                    example: 123
                  your_score:
                    type: integer
                    example: 2500
                  all_finished:
                    type: boolean
                    description: 모든 참가자 점수 제출 여부
                    example: true
                  message:
                    type: string
                    example: "게임이 종료되었습니다"
                  final_scores:
                    type: object
                    description: 참가자별 최종 점수 (모든 참가자 제출 시)
                    additionalProperties:
                      type: integer
                    example:
                      user123: 2500
                      user456: 1800
                  winner:
                    type: string
                    description: 승자 이름 (모든 참가자 제출 시)
                    example: "홍길동"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # 랭킹 API
  /api/ranking/score:
    get:
      tags:
        - Ranking
      summary: 점수 랭킹 조회
      description: 전체 사용자 중 최고 점수 TOP 100을 조회합니다
      security: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
          description: 조회할 랭킹 수 (기본값 100)
      responses:
        "200":
          description: 점수 랭킹 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  rankings:
                    type: array
                    items:
                      type: object
                      properties:
                        rank:
                          type: integer
                          description: 순위
                          example: 1
                        user_name:
                          type: string
                          description: 사용자명
                          example: "홍길동"
                        score:
                          type: integer
                          description: 최고 점수
                          example: 5000
                        played_at:
                          type: string
                          format: date-time
                          description: 달성 일시
                          example: "2025-07-08T14:30:00Z"
                    example:
                      - rank: 1
                        user_name: "홍길동"
                        score: 5000
                        played_at: "2025-07-08T14:30:00Z"
                      - rank: 2
                        user_name: "김철수"
                        score: 4800
                        played_at: "2025-07-08T13:15:00Z"
                      - rank: 3
                        user_name: "이영희"
                        score: 4200
                        played_at: "2025-07-08T12:45:00Z"
                  total_count:
                    type: integer
                    description: 전체 랭킹 수
                    example: 50

  /api/ranking/wins:
    get:
      tags:
        - Ranking
      summary: 승리 횟수 랭킹 조회
      description: 대전 모드 승리 횟수 TOP 100을 조회합니다
      security: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
          description: 조회할 랭킹 수 (기본값 100)
      responses:
        "200":
          description: 승리 횟수 랭킹 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  rankings:
                    type: array
                    items:
                      type: object
                      properties:
                        rank:
                          type: integer
                          description: 순위
                          example: 1
                        user_name:
                          type: string
                          description: 사용자명
                          example: "김철수"
                        wins:
                          type: integer
                          description: 승리 횟수
                          example: 25
                    example:
                      - rank: 1
                        user_name: "김철수"
                        wins: 25
                      - rank: 2
                        user_name: "홍길동"
                        wins: 22
                      - rank: 3
                        user_name: "이영희"
                        wins: 18
                  total_count:
                    type: integer
                    description: 전체 랭킹 수
                    example: 30

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "유효성 검증 실패"
              message:
                type: string
                example: "입력 데이터를 확인해주세요"
              details:
                type: array
                items:
                  type: string
                example:
                  ["아이디는 5-20자여야 합니다", "비밀번호가 일치하지 않습니다"]

    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "인증 실패"
              message:
                type: string
                example: "유효하지 않은 토큰입니다"

    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "권한 없음"
              message:
                type: string
                example: "방장만 이 작업을 수행할 수 있습니다"

    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "리소스 없음"
              message:
                type: string
                example: "요청한 방을 찾을 수 없습니다"

    Conflict:
      description: 중복된 리소스
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "중복 오류"
              message:
                type: string
                example: "이미 사용 중인 아이디입니다"

    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "서버 오류"
              message:
                type: string
                example: "서버에서 오류가 발생했습니다. 잠시 후 다시 시도해주세요"
  tags:
    - name: Rooms
      description: 방 관리 관련 API
    - name: Game
      description: 게임 플레이 관련 API
    - name: Ranking
      description: 랭킹 조회 관련 API